refactor: move painting type
